requirements.txt
Flask==2.2.2
gunicorn==20.1.0
psycopg2-binary==2.9.5
Flask-Cors==3.0.10